TIP-842: Move association type model to structure sub context
