+++ src/Akeneo/Pim/Structure/Bundle/Resources/config/entities.yml
@@ -14,3 +14,16 @@ parameters:
     pim_catalog.model.group_type_translation.interface: Akeneo\Pim\Structure\Component\Model\GroupTypeTranslationInterface
     pim_catalog.entity.group_type.class: Akeneo\Pim\Structure\Component\Model\GroupType
     pim_catalog.entity.group_type_translation.class: Akeneo\Pim\Structure\Component\Model\GroupTypeTranslation
+    pim_catalog.model.attribute.interface: Akeneo\Pim\Structure\Component\Model\AttributeInterface
+    pim_catalog.model.attribute_group.interface: Akeneo\Pim\Structure\Component\Model\AttributeGroupInterface
+    pim_catalog.model.attribute_option.interface: Akeneo\Pim\Structure\Component\Model\AttributeOptionInterface
+    pim_catalog.model.attribute_requirement.interface: Akeneo\Pim\Structure\Component\Model\AttributeRequirementInterface
+    pim_catalog.model.family.interface: Akeneo\Pim\Structure\Component\Model\FamilyInterface
+    pim_catalog.model.family_variant.interface: Akeneo\Pim\Structure\Component\Model\FamilyVariantInterface
+    pim_catalog.model.family_translation.interface: Akeneo\Pim\Structure\Component\Model\FamilyTranslationInterface
+    pim_catalog.model.attribute_group_translation.interface: Akeneo\Pim\Structure\Component\Model\AttributeGroupTranslationInterface
+    pim_catalog.model.attribute_translation.interface: Akeneo\Pim\Structure\Component\Model\AttributeTranslationInterface
+    pim_catalog.model.association_type.interface: Akeneo\Pim\Structure\Component\Model\AssociationTypeInterface
+    pim_catalog.entity.association_type.class: Akeneo\Pim\Structure\Component\Model\AssociationType
+    pim_catalog.entity.association_type_translation.class: Akeneo\Pim\Structure\Component\Model\AssociationTypeTranslation
+    pim_catalog.model.association_type_translation.interface: Akeneo\Pim\Structure\Component\Model\AssociationTypeTranslationInterface
