+++ src/Akeneo/Pim/Structure/Component/Model/AssociationTypeInterface.php
@@ -1,10 +1,11 @@
 <?php
 
-namespace Pim\Component\Catalog\Model;
+namespace Akeneo\Pim\Structure\Component\Model;
 
 use Akeneo\Tool\Component\Localization\Model\TranslatableInterface;
 use Akeneo\Tool\Component\StorageUtils\Model\ReferableInterface;
 use Akeneo\Tool\Component\Versioning\Model\VersionableInterface;
+use Pim\Component\Catalog\Model\TimestampableInterface;
 
 /**
  * Association type
