+++ src/Akeneo/Tool/Component/Api/Normalizer/AssociationTypeNormalizer.php
@@ -2,7 +2,7 @@
 
 namespace Akeneo\Tool\Component\Api\Normalizer;
 
-use Pim\Component\Catalog\Model\AssociationTypeInterface;
+use Akeneo\Pim\Structure\Component\Model\AssociationTypeInterface;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 
 /**
