+++ src/Pim/Bundle/CatalogBundle/Resources/config/entities.yml
@@ -6,8 +6,6 @@ parameters:
     pim_catalog.entity.association.class:                  Pim\Component\Catalog\Model\ProductAssociation
     pim_catalog.entity.product_model_association.class:    Pim\Component\Catalog\Model\ProductModelAssociation
     pim_catalog.entity.completeness.class:                 Pim\Component\Catalog\Model\Completeness
-    pim_catalog.entity.association_type.class:             Pim\Bundle\CatalogBundle\Entity\AssociationType
-    pim_catalog.entity.association_type_translation.class: Pim\Bundle\CatalogBundle\Entity\AssociationTypeTranslation
     pim_catalog.entity.category.class:                     Pim\Bundle\CatalogBundle\Entity\Category
     pim_catalog.entity.category_translation.class:         Pim\Bundle\CatalogBundle\Entity\CategoryTranslation
     pim_catalog.entity.group.class:                        Pim\Bundle\CatalogBundle\Entity\Group
