+++ src/Pim/Bundle/CatalogBundle/Resources/config/model/doctrine/ProductModelAssociation.orm.yml
@@ -15,7 +15,7 @@ Pim\Component\Catalog\Model\ProductModelAssociation:
                 - association_type_id
     manyToOne:
         associationType:
-            targetEntity: Pim\Component\Catalog\Model\AssociationTypeInterface
+            targetEntity: Akeneo\Pim\Structure\Component\Model\AssociationTypeInterface
             joinColumns:
                 association_type_id:
                     referencedColumnName: id
