+++ src/Pim/Bundle/DataGridBundle/spec/Extension/MassAction/Util/ProductFieldsBuilderSpec.php
@@ -4,7 +4,7 @@ namespace spec\Pim\Bundle\DataGridBundle\Extension\MassAction\Util;
 
 use PhpSpec\ObjectBehavior;
 use Pim\Bundle\CatalogBundle\Context\CatalogContext;
-use Pim\Component\Catalog\Model\AssociationTypeInterface;
+use Akeneo\Pim\Structure\Component\Model\AssociationTypeInterface;
 use Akeneo\Pim\Structure\Component\Model\AttributeInterface;
 use Pim\Component\Catalog\Repository\AssociationTypeRepositoryInterface;
 use Akeneo\Pim\Structure\Component\Repository\AttributeRepositoryInterface;
