+++ src/Pim/Bundle/EnrichBundle/spec/Form/Type/AssociationTypeSpec.php
@@ -4,7 +4,7 @@ namespace spec\Pim\Bundle\EnrichBundle\Form\Type;
 
 use Doctrine\ORM\EntityManager;
 use PhpSpec\ObjectBehavior;
-use Pim\Bundle\CatalogBundle\Entity\AssociationType;
+use Akeneo\Pim\Structure\Component\Model\AssociationType;
 use Pim\Bundle\CatalogBundle\Entity\Group;
 use Pim\Bundle\EnrichBundle\Form\Type\ObjectIdentifierType;
 use Pim\Component\Catalog\Model\AssociationInterface;
