+++ src/Pim/Bundle/VersioningBundle/spec/Normalizer/Flat/ProductModelNormalizerSpec.php
@@ -7,7 +7,7 @@ namespace spec\Pim\Bundle\VersioningBundle\Normalizer\Flat;
 use Doctrine\Common\Collections\ArrayCollection;
 use PhpSpec\ObjectBehavior;
 use Pim\Bundle\VersioningBundle\Normalizer\Flat\ProductModelNormalizer;
-use Pim\Component\Catalog\Model\AssociationTypeInterface;
+use Akeneo\Pim\Structure\Component\Model\AssociationTypeInterface;
 use Akeneo\Pim\Structure\Component\Model\FamilyVariantInterface;
 use Pim\Component\Catalog\Model\GroupInterface;
 use Pim\Component\Catalog\Model\ProductInterface;
