+++ src/Pim/Component/Catalog/Repository/AssociationTypeRepositoryInterface.php
@@ -4,7 +4,7 @@ namespace Pim\Component\Catalog\Repository;
 
 use Akeneo\Tool\Component\StorageUtils\Repository\IdentifiableObjectRepositoryInterface;
 use Doctrine\Common\Persistence\ObjectRepository;
-use Pim\Component\Catalog\Model\AssociationTypeInterface;
+use Akeneo\Pim\Structure\Component\Model\AssociationTypeInterface;
 use Pim\Component\Catalog\Model\EntityWithAssociationsInterface;
 
 /**
