+++ src/Pim/Component/Catalog/spec/Normalizer/Standard/AssociationTypeNormalizerSpec.php
@@ -3,7 +3,7 @@
 namespace spec\Pim\Component\Catalog\Normalizer\Standard;
 
 use PhpSpec\ObjectBehavior;
-use Pim\Component\Catalog\Model\AssociationTypeInterface;
+use Akeneo\Pim\Structure\Component\Model\AssociationTypeInterface;
 use Pim\Component\Catalog\Normalizer\Standard\TranslationNormalizer;
 
 class AssociationTypeNormalizerSpec extends ObjectBehavior
