+++ src/Pim/Component/Connector/spec/ArrayConverter/FlatToStandard/Product/AttributeColumnInfoExtractorSpec.php
@@ -11,7 +11,7 @@ use Pim\Component\Connector\ArrayConverter\FlatToStandard\Product\AssociationCol
 
 class AttributeColumnInfoExtractorSpec extends ObjectBehavior
 {
-    const ASSOC_TYPE_CLASS = 'Pim\Bundle\CatalogBundle\Entity\AssociationType';
+    const ASSOC_TYPE_CLASS = 'Akeneo\Pim\Structure\Component\Model\AssociationType';
     const ATTRIBUTE_CLASS = 'Akeneo\Pim\Structure\Component\Model\Attribute';
     const CHANNEL_CLASS = 'Akeneo\Channel\Component\Model\Channel';
     const LOCALE_CLASS = 'Akeneo\Channel\Component\Model\Locale';
